Parse multiboot2 info (wip)

+++ kernel/src/init/mod.rs
@@ -3,7 +3,7 @@ pub mod boot;
 pub mod color;
 pub mod cpu;
 
-pub fn init(vga_index: &mut isize, multiboot_ptr: u32) {
+pub fn init(vga_index: &mut isize, multiboot_ptr: u64) {
     ascii::ascii_art(vga_index);
     cpu::print_mode(vga_index);
     boot::print_info(vga_index, multiboot_ptr);
